Move components into 'utilities' directory.
