+++ resources/assets/components/utilities/PostCard/post.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .post {
   font-family: $primary-font-family;
